crud, banda y representante

+++ database/migrations/2023_06_07_054034_create_detalle_representante_table.php
@@ -17,7 +17,7 @@ class CreateDetalleRepresentanteTable extends Migration
             $table->increments('id');
             $table->string('nombre');
             $table->string('email');
-            $table->integer('contacto');
+            $table->bigInteger('contacto');
             $table->timestamps();
         });
     }
